main: split main.go into more files
License: MIT
