+++ cmd/ipfs/debug.go
@@ -1,0 +1,75 @@
+package main
+
+import (
+	"os"
+	"runtime/pprof"
+	"time"
+
+	cmds "github.com/ipfs/go-ipfs-cmds"
+	u "github.com/ipfs/go-ipfs-util"
+	logging "github.com/ipfs/go-log"
+)
+
+
+func checkDebug(req *cmds.Request) {
+	// check if user wants to debug. option OR env var.
+	debug, _ := req.Options["debug"].(bool)
+	if debug || os.Getenv("IPFS_LOGGING") == "debug" {
+		u.Debug = true
+		logging.SetDebugLogging()
+	}
+	if u.GetenvBool("DEBUG") {
+		u.Debug = true
+	}
+}
+
+// startProfiling begins CPU profiling and returns a `stop` function to be
+// executed as late as possible. The stop function captures the memprofile.
+func startProfiling() (func(), error) {
+	// start CPU profiling as early as possible
+	ofi, err := os.Create(cpuProfile)
+	if err != nil {
+		return nil, err
+	}
+	err = pprof.StartCPUProfile(ofi)
+	if err != nil {
+		ofi.Close()
+		return nil, err
+	}
+	go func() {
+		for range time.NewTicker(time.Second * 30).C {
+			err := writeHeapProfileToFile()
+			if err != nil {
+				log.Error(err)
+			}
+		}
+	}()
+
+	stopProfiling := func() {
+		pprof.StopCPUProfile()
+		ofi.Close() // captured by the closure
+	}
+	return stopProfiling, nil
+}
+
+func writeHeapProfileToFile() error {
+	mprof, err := os.Create(heapProfile)
+	if err != nil {
+		return err
+	}
+	defer mprof.Close() // _after_ writing the heap profile
+	return pprof.WriteHeapProfile(mprof)
+}
+
+func profileIfEnabled() (func(), error) {
+	// FIXME this is a temporary hack so profiling of asynchronous operations
+	// works as intended.
+	if os.Getenv(EnvEnableProfiling) != "" {
+		stopProfilingFunc, err := startProfiling() // TODO maybe change this to its own option... profiling makes it slower.
+		if err != nil {
+			return nil, err
+		}
+		return stopProfilingFunc, nil
+	}
+	return func() {}, nil
+}
